Rename actions and files to include data development focus
This also improves the tests to check that only dmd codes in the codelists are selected.
