+++ project.yaml
@@ -5,10 +5,10 @@ expectations:
   population_size: 1000
 
 actions:
-  generate_dataset_curation:
+  generate_pf_data_development:
     run: >
-      ehrql:v1 generate-dataset analysis/dataset_definition_curation.py
-        --test-data-file analysis/test_dataset_definition_curation.py
+      ehrql:v1 generate-dataset analysis/dataset_definition_pf_data_development.py
+        --test-data-file analysis/test_dataset_definition_pf_data_development.py
         --dummy-tables dummy_tables
         --output output/data_curation/data_curation.csv.gz
     outputs:
@@ -26,14 +26,14 @@ actions:
 
   data_curation_medications:
     run: r:latest analysis/data_curation_medications_table.R
-    needs: [generate_dataset_curation]
+    needs: [generate_pf_data_development]
     outputs:
       moderately_sensitive:
         dataset: output/data_curation/medication_status_counts.csv
 
   data_curation_consultations:
     run: r:latest analysis/data_curation_pharmacy_first_consultations.R
-    needs: [generate_dataset_curation]
+    needs: [generate_pf_data_development]
     outputs:
       moderately_sensitive:
         dataset: output/data_curation/consultations_counts.csv
